coverted the bot to modular structure
