+++ src/core/config.py
@@ -1,0 +1,86 @@
+"""
+Configuration module for the Telegram File Downloader Bot.
+Handles environment variables, settings, and validation.
+"""
+
+import os
+import logging
+import pathlib
+from dotenv import load_dotenv
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+    handlers=[
+        logging.StreamHandler()
+    ]
+)
+logger = logging.getLogger(__name__)
+
+class Config:
+    """Configuration class for the bot."""
+    
+    def __init__(self):
+        # Load environment variables
+        load_dotenv()
+        
+        # Debug: Check if .env file exists and what's loaded
+        env_path = pathlib.Path('.env')
+        logger.info(f".env file exists: {env_path.exists()}")
+        logger.info(f"Current working directory: {pathlib.Path.cwd()}")
+        
+        # Check what's actually loaded
+        logger.info(f"API_ID from env: {repr(os.getenv('API_ID'))}")
+        logger.info(f"API_HASH from env: {repr(os.getenv('API_HASH'))}")
+        logger.info(f"BOT_TOKEN from env: {repr(os.getenv('BOT_TOKEN'))}")
+        logger.info(f"ALLOWED_USERS from env: {repr(os.getenv('ALLOWED_USERS'))}")
+        
+        # Load configuration
+        self.api_id = os.getenv('API_ID')
+        self.api_hash = os.getenv('API_HASH')
+        self.bot_token = os.getenv('BOT_TOKEN')
+        self.allowed_users = set(os.getenv('ALLOWED_USERS', '').split(','))
+        
+        # Bot settings
+        self.session_name = 'downloader_bot_session'
+        self.base_download_dir = '.'
+        
+        # Download settings
+        self.download_chunk_size = 1024 * 1024  # 1MB chunks
+        self.progress_update_interval = 5.0  # Update progress every 5 seconds
+        self.max_concurrent_downloads = 5
+        self.notification_cooldown = 30  # Minimum seconds between notifications per user
+        
+        # Connection settings
+        self.connection_retries = 5
+        self.retry_delay = 1
+        self.timeout = 30
+        self.request_retries = 5
+        self.flood_sleep_threshold = 60
+        
+        self._validate_config()
+        self._log_config()
+    
+    def _validate_config(self):
+        """Validate required configuration values."""
+        if not self.api_id or not self.api_hash or not self.bot_token:
+            logger.error("Missing required environment variables")
+            raise ValueError('API_ID, API_HASH, and BOT_TOKEN must be set in the environment variables.')
+        
+        # Convert API_ID to int
+        try:
+            self.api_id = int(self.api_id)
+        except ValueError:
+            raise ValueError('API_ID must be a valid integer.')
+        
+        logger.info(f"API_ID converted to int: {self.api_id}")
+    
+    def _log_config(self):
+        """Log configuration details (without sensitive data)."""
+        logger.info(f"Loaded environment variables - API_ID: {self.api_id}, API_HASH: {'*' * len(self.api_hash) if self.api_hash else 'None'}, BOT_TOKEN: {'*' * len(self.bot_token) if self.bot_token else 'None'}")
+        logger.info(f"Allowed users: {self.allowed_users}")
+        logger.info("Configuration validation completed successfully")
+
+# Global configuration instance
+config = Config()
